Initial implementation of SysMLTransientValueService
We must mark attributes and references derived from the parsed model as
transients, so that Xtext will not attempt to serialize them.

Currently, the SysMLTransientValueService only marks transient values
appearing in the VehicleDefinitions.sysml example as transient. To
support serialization of further language elements, it should be
extended in the future.